+++ django/sources/umamagic/urls.py
@@ -21,5 +21,4 @@ from django.urls import include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('event_schedules/', include('apps.event_schedules.urls')),
-    path('web_controller/', include('apps.web_controller.urls')),
 ]
